cleanup and put BlockSums back in chain crate
